+++ src/graphql/account_details_documents.ts
@@ -195,3 +195,35 @@ query StakeAccountDetails($address: String) {
   }
 }
 `;
+
+// vote detail account
+export const VoteAccountDetailsDocument = /* GraphQL */`
+query VoteAccountDetails($address: String) {
+  validator(where: {address: {_eq: $address}}) {
+    address
+    nativeBalance: native_balance {
+      balance
+    }
+    commission
+    voter
+    withdrawer
+    validatorStatus: validator_status {
+      rootSlot: root_slot
+      lastVote: last_vote
+      active
+    }
+    validatorConfig: validator_config {
+      details
+      avatarUrl: avatar_url
+      name
+      website
+    }
+    validatorSkipRate: validator_skip_rate {
+      epoch
+      skip
+      skipRate: skip_rate
+      total
+    }
+  }
+}
+`;