+++ src/screens/account_details/components/native_account/index.tsx
@@ -1,15 +1,11 @@
 import React from 'react';
 import {
-  LoadAndExist,
-} from '@components';
-import {
   Overview,
   Balance,
   Accounts,
 } from './components';
 import {
   Tokens,
-  Transactions,
 } from '..';
 import { useNativeAccount } from './hooks';
 import { useStyles } from './styles';
@@ -19,16 +15,12 @@ const NativeAccount = () => {
   const { state } = useNativeAccount();
 
   return (
-    <LoadAndExist
-      loading={state.loading}
-      exists={state.exists}
-    >
+    <>
       <Overview className={classes.overview} overview={state.overview} />
       <Balance className={classes.balance} balance={state.balance} />
       <Accounts className={classes.accounts} />
       <Tokens className={classes.tokens} />
-      <Transactions className={classes.transactions} />
-    </LoadAndExist>
+    </>
   );
 };