+++ src/screens/account_details/components/token_account/index.tsx
@@ -1,14 +1,10 @@
 import React from 'react';
 import {
-  LoadAndExist,
-} from '@components';
-import {
   Overview,
   Market,
   Header,
   TopHolders,
 } from './components';
-import { Transactions } from '..';
 import { useStyles } from './styles';
 import { useTokenAccount } from './hooks';
 
@@ -18,25 +14,18 @@ const TokenAccount = () => {
     state,
   } = useTokenAccount();
   return (
-    <LoadAndExist
-      loading={state.loading}
-      exists={state.exists}
-    >
-      <div className={classes.root}>
-        <Header className={classes.header} header={state.header} />
-        <Overview
-          className={classes.overview}
-          overview={state.overview}
-        />
-        <Market
-          className={classes.market}
-          market={state.market}
-        />
-        {/* <Holders className={classes.holders} /> */}
-        <TopHolders className={classes.holders} />
-        <Transactions className={classes.transactions} />
-      </div>
-    </LoadAndExist>
+    <div className={classes.root}>
+      <Header className={classes.header} header={state.header} />
+      <Overview
+        className={classes.overview}
+        overview={state.overview}
+      />
+      <Market
+        className={classes.market}
+        market={state.market}
+      />
+      <TopHolders className={classes.holders} />
+    </div>
   );
 };