+++ src/screens/account_details/components/vote_account/components/overview/types.ts
@@ -1,9 +1,0 @@
-export type OverviewType = {
-  address: string;
-  balance: string;
-  authorizedVoter: string;
-  authorizedWithdrawer: string;
-  lastTimestamp: string;
-  commission: number;
-  rootSlot: number;
-}