+++ CHANGELOG.md
@@ -29,4 +29,5 @@
 - Add nonce account details page
 - Add token details account page
 - Add stake details account page
+- Add vote details account page